+++ src/redux/appConfig/appConfig.slice.ts
@@ -10,6 +10,7 @@ import {CONFIG_PATH, keysToDelete, keysToUpdateStateBulk, writeProjectConfigFile
 import {setRootFolder} from '@redux/thunks/setRootFolder';
 
 import {init as sentryInit} from '@sentry/electron/renderer';
+import {SKIPPED_ERRORS} from '@shared/constants/sentry';
 import {ClusterColors} from '@shared/models/cluster';
 import {
   ApiKeyVendor,
@@ -370,9 +371,8 @@ export const configSlice = createSlice({
           dsn: action.payload.SENTRY_DSN,
           tracesSampleRate: 0.6,
           beforeSend: event => {
-            // TODO: Skip errors related to model.getModeId for now, should fix in 2.0
-            // also why is it model.getModeId and not model.getModelId? is it a typo?
-            if (event.exception?.values?.some(v => v.value?.includes('model.getModeId'))) {
+            // skipping noisy errors from sentry
+            if (event.exception?.values?.some(v => SKIPPED_ERRORS.some(err => v.value?.includes(err)))) {
               return null;
             }
             // we have to get this from electron store to get the most updated value